+++ history.md
@@ -1,3 +1,12 @@
+* Fixed combat reload crash
+* Relationship met_before property now set correctly
+* Portrait will appear behind info box in skill trainer
+* Monster unit minimum monster price increased
+* Fixed Spot Weakness crash
+* get_odds defense method renamed to get_defense_odds
+* Missiles more likely to hit and penetrate
+* Added anim for Spot Weakness
+
 v1.006alpha March 4, 2026
 * ItemStatsBlock will now deal correctly with linewrap
 * Added styles for MenuWidget; new default style